Move 'cleanAmount' to formatting library plus 'isNumeric' to input
